Rename .js files back to .jsx
