+++ index.js
@@ -1,2 +1,2 @@
 
-module.exports = require('./src/js/example.js').default;
+module.exports = require('./src/js/example.jsx').default;
