+++ src/js/bar.jsx
